Moved node id input into NodeIdInput.tsx
