+++ src/components/sidebar/nodeView/NodeForm.tsx
@@ -2,22 +2,20 @@ import classnames from 'classnames';
 import _ from 'lodash';
 import { inject, observer } from 'mobx-react';
 import React, { Component } from 'react';
-import { Dropdown, RadioButton, TransitToggleButtonBar } from '~/components/controls';
+import { Dropdown } from '~/components/controls';
 import { IDropdownItem } from '~/components/controls/Dropdown';
 import InputContainer from '~/components/controls/InputContainer';
 import TextContainer from '~/components/controls/TextContainer';
 import NodeMeasurementType from '~/enums/nodeMeasurementType';
 import NodeType from '~/enums/nodeType';
-import TransitType from '~/enums/transitType';
 import { INode } from '~/models';
-import NodeService from '~/services/nodeService';
 import { CodeListStore } from '~/stores/codeListStore';
 import { NodeStore } from '~/stores/nodeStore';
 import NodeLocationType from '~/types/NodeLocationType';
 import NodeUtils from '~/utils/NodeUtils';
 import { createDropdownItemsFromList } from '~/utils/dropdownUtils';
-import { IValidationResult } from '~/validation/FormValidator';
 import CoordinateInputRow from './CoordinateInputRow';
+import NodeIdInput from './NodeIdInput';
 import * as s from './nodeForm.scss';
 
 interface INodeFormProps {
@@ -33,39 +31,9 @@ interface INodeFormProps {
     codeListStore?: CodeListStore;
 }
 
-interface INodeFormState {
-    nodeIdSuffixOptions: IDropdownItem[];
-    isNodeIdSuffixQueryLoading: boolean;
-    beginningOfNodeIdValidationResult: IValidationResult;
-    idSuffixValidationResult: IValidationResult;
-}
-
 @inject('nodeStore', 'codeListStore')
 @observer
-class NodeForm extends Component<INodeFormProps, INodeFormState> {
-    state = {
-        nodeIdSuffixOptions: [],
-        isNodeIdSuffixQueryLoading: false,
-        beginningOfNodeIdValidationResult: { isValid: false },
-        idSuffixValidationResult: { isValid: false },
-    };
-    private _isMounted: boolean;
-    private _setState = (newState: object) => {
-        if (this._isMounted) {
-            this.setState(newState);
-        }
-    };
-    componentDidMount() {
-        this._isMounted = true;
-        if (!this.props.isEditingDisabled && this.props.nodeStore!.node.beginningOfNodeId) {
-            this.queryAvailableNodeIdSuffixes();
-        }
-    }
-
-    componentWillUnmount() {
-        this._isMounted = false;
-    }
-
+class NodeForm extends Component<INodeFormProps> {
     private createMeasuredDropdownItems = (): IDropdownItem[] => {
         const items: IDropdownItem[] = [
             {
@@ -80,93 +48,6 @@ class NodeForm extends Component<INodeFormProps, INodeFormState> {
         return items;
     };
 
-    private onChangeNodeId = async (beginningOfNodeId: string) => {
-        this.props.onChangeNodeProperty!('beginningOfNodeId')(beginningOfNodeId);
-
-        if (beginningOfNodeId.length === 4) {
-            await this.queryAvailableNodeIdSuffixes();
-        } else {
-            this._setState({
-                nodeIdSuffixOptions: [],
-            });
-            this.props.onChangeNodeProperty!('idSuffix')(null);
-        }
-    };
-
-    private selectTransitType = async (transitType: TransitType) => {
-        this._setState({
-            nodeIdSuffixOptions: [],
-        });
-        this.props.onChangeNodeProperty!('transitType')(transitType);
-        this.props.onChangeNodeProperty!('idSuffix')(null);
-        this.props.onChangeNodeProperty!('isInternal')(false);
-        await this.queryAvailableNodeIdSuffixes();
-    };
-
-    private queryAvailableNodeIdSuffixes = async () => {
-        const nodeStore = this.props.nodeStore!;
-        const beginningOfNodeId = nodeStore.node.beginningOfNodeId;
-        if (!beginningOfNodeId || beginningOfNodeId.length !== 4) return;
-
-        this._setState({
-            isNodeIdSuffixQueryLoading: true,
-        });
-
-        const nodeIdUsageCode = this.getNodeIdUsageCode();
-        const availableNodeIds = await NodeService.fetchAvailableNodeIdsWithPrefix(
-            `${beginningOfNodeId}${nodeIdUsageCode}`
-        );
-
-        // slide(-2): get last two letters of a nodeId
-        const nodeIdSuffixList = availableNodeIds.map((nodeId: string) => nodeId.slice(-2));
-
-        this._setState({
-            nodeIdSuffixOptions: createDropdownItemsFromList(nodeIdSuffixList),
-            isNodeIdSuffixQueryLoading: false,
-        });
-    };
-
-    private getNodeIdUsageCode = () => {
-        const nodeStore = this.props.nodeStore!;
-        const transitType = nodeStore.node.transitType;
-        const isInternal = nodeStore.node.isInternal;
-        switch (transitType) {
-            case TransitType.BUS:
-                if (isInternal) {
-                    // Helsinki
-                    return '1'; // Hki internal network
-                }
-                return '2'; // Regional bus network
-            case TransitType.TRAM:
-                return '4';
-            case TransitType.TRAIN:
-                return '5';
-            case TransitType.SUBWAY:
-                return '6';
-            case TransitType.FERRY:
-                return '7';
-            default:
-                return '3'; // Default number that is not restricted to any usage
-        }
-    };
-
-    private onChangeNodeIdSuffix = (idSuffix: string) => {
-        const node = this.props.nodeStore!.node;
-        const nodeIdUsageCode = this.getNodeIdUsageCode();
-        const nodeId = `${node.beginningOfNodeId}${nodeIdUsageCode}${idSuffix}`;
-        this.props.onChangeNodeProperty!('id')(nodeId);
-        this.props.onChangeNodeProperty!('idSuffix')(idSuffix);
-    };
-
-    private onChangeIsInternal = (isInternal: boolean) => async () => {
-        this._setState({
-            isInternal,
-        });
-        this.props.onChangeNodeProperty!('isInternal')(isInternal);
-        this.props.onChangeNodeProperty!('idSuffix')(null);
-        await this.queryAvailableNodeIdSuffixes();
-    };
-
     render() {
         const {
             node,
@@ -179,69 +60,17 @@ class NodeForm extends Component<INodeFormProps, INodeFormState> {
             onChangeNodeType,
         } = this.props;
         const nodeTypeCodeList = createDropdownItemsFromList(['P', 'X']);
-        const { nodeIdSuffixOptions, isNodeIdSuffixQueryLoading } = this.state;
         return (
             <div className={classnames(s.nodeForm, s.form)}>
                 <div className={s.formSection}>
                     {isNewNode && (
-                        <>
-                            <div className={s.flexRow}>
-                                <InputContainer
-                                    value={isNodeIdEditable ? node.beginningOfNodeId : node.id}
-                                    onChange={this.onChangeNodeId}
-                                    label={
-                                        isNodeIdEditable ? 'SOLMUN TUNNUS (4 num.' : 'SOLMUN TUNNUS'
-                                    }
-                                    disabled={
-                                        !isNodeIdEditable || Boolean(isNodeIdSuffixQueryLoading)
-                                    }
-                                    validationResult={invalidPropertiesMap['beginningOfNodeId']}
-                                    data-cy='nodeId'
-                                />
-                                {isNodeIdEditable && (
-                                    <Dropdown
-                                        label='+ 2 num.)'
-                                        onChange={this.onChangeNodeIdSuffix}
-                                        disabled={_.isEmpty(nodeIdSuffixOptions)}
-                                        isLoading={isNodeIdSuffixQueryLoading}
-                                        selected={node.idSuffix}
-                                        items={nodeIdSuffixOptions ? nodeIdSuffixOptions : []}
-                                        validationResult={invalidPropertiesMap['idSuffix']}
-                                        data-cy='idSuffix'
-                                    />
-                                )}
-                            </div>
-                            <div className={s.flexRow}>
-                                <div className={s.formItem}>
-                                    <div className={s.inputLabel}>VERKKO</div>
-                                    <TransitToggleButtonBar
-                                        selectedTransitTypes={
-                                            node.transitType ? [node.transitType!] : []
-                                        }
-                                        toggleSelectedTransitType={this.selectTransitType}
-                                        errorMessage={''}
-                                    />
-                                </div>
-                            </div>
-                            {node.transitType && node.transitType === '1' && (
-                                <div className={s.flexRow}>
-                                    <div className={s.formItem}>
-                                        <RadioButton
-                                            onClick={this.onChangeIsInternal(true)}
-                                            checked={Boolean(node.isInternal)}
-                                            text={'Helsingin sisäinen'}
-                                        />
-                                    </div>
-                                    <div className={s.formItem}>
-                                        <RadioButton
-                                            onClick={this.onChangeIsInternal(false)}
-                                            checked={Boolean(!node.isInternal)}
-                                            text={'Helsingin ulkopuolinen'}
-                                        />
-                                    </div>
-                                </div>
-                            )}
-                        </>
+                        <NodeIdInput
+                            node={node}
+                            isEditingDisabled={isEditingDisabled}
+                            invalidPropertiesMap={invalidPropertiesMap}
+                            isNodeIdEditable={isNodeIdEditable}
+                            onChangeNodeProperty={onChangeNodeProperty}
+                        />
                     )}
                     <div className={s.flexRow}>
                         <Dropdown
